+++ src/mcresponse/messages.nim
@@ -9,129 +9,139 @@
 #
 
 import json, httpcore, tables
+import netcode
 
 ## Type definition for the stardard responses / messages
 type
     ResponseMessage* = object
-        code*: string
+        code*: MessageCode
         resCode*: HttpCode
         message*: string
         value*: JsonNode
     
 ## Standard default responses: includes app-code, http-code, message and transaction value (json-object)
 ## 
-var responseMessages* = initTable[string, ResponseMessage]()
+var responseMessages* = initTable[MessageCode, ResponseMessage]()
 
-responseMessages["checkEror"] = ResponseMessage(
-        code   : "paramsError",
-        resCode: Http406,
+responseMessages = {
+        MessageCode.CheckErrorCode: ResponseMessage(
+        code   : MessageCode.ParamErrorCode,
+        resCode: StatusCodeNotAcceptable,
         message: "Parameters checking error",
-        value  : JsonNode(nil) )
-
-responseMessages["connectError"] = ResponseMessage(
-        code   : "connectionError",
-        resCode: Http503,
+        value  : JsonNode(nil) ),
+        
+        MessageCode.ConnectionErrorCode: ResponseMessage(
+        code   : MessageCode.ConnectionErrorCode,
+        resCode: StatusCodeServiceUnavailable,
         message: "Connection error",
-        value  : JsonNode(nil) )
+        value  : JsonNode(nil) ),
 
-responseMessages["validateError"] = ResponseMessage(
-        code   : "paramsError",
-        resCode: Http406,
+        MessageCode.ValidateErrorCode: ResponseMessage(
+        code   : MessageCode.ParamErrorCode,
+        resCode: StatusCodeNotAcceptable,
         message: "Validation error for inputs parameters",
-        value  : JsonNode(nil) )
+        value  : JsonNode(nil) ),
 
-responseMessages["tokenExpired"] = ResponseMessage(
-        code   : "tokenExpired",
-        resCode: Http401,
+        MessageCode.TokenExpiredCode: ResponseMessage(
+        code   : MessageCode.TokenExpiredCode,
+        resCode: StatusCodeUnAuthorized,
         message: "Unauthorized. Token / Access-key has expired. Please login again",
-        value  : JsonNode(nil) )
+        value  : JsonNode(nil) ),
 
-responseMessages["unAuthorized"] = ResponseMessage(
-        code   : "unAuthorized",
-        resCode: Http401,
+        MessageCode.UnAuthorizedCode: ResponseMessage(
+        code   : MessageCode.UnAuthorizedCode,
+        resCode: StatusCodeUnAuthorized,
         message: "Unauthorised Action or Task",
-        value  : JsonNode(nil) )
-
-responseMessages["notFound"] = ResponseMessage(
-        code   : "notFound",
-        resCode: Http404,
+        value  : JsonNode(nil) ),
+        
+        MessageCode.NotFoundCode: ResponseMessage(
+        code   :MessageCode.NotFoundCode,
+        resCode: StatusCodeNotFound,
         message: "Requested information not found",
-        value  : JsonNode(nil) )
+        value  : JsonNode(nil) ),
 
-responseMessages["success"] = ResponseMessage(
-        code   : "success",
-        resCode: Http200,
+        MessageCode.SuccessCode: ResponseMessage(
+        code   : MessageCode.SuccessCode,
+        resCode: StatusCodeOK,
         message: "Request completed successfully",
-        value  : JsonNode(nil) )
+        value  : JsonNode(nil) ),
 
-responseMessages["removeDenied"] = ResponseMessage(
-        code   : "removeDenied",
-        resCode: Http401,
+        MessageCode.RemoveDeniedCode: ResponseMessage(
+        code   : MessageCode.RemoveDeniedCode,
+        resCode: StatusCodeUnAuthorized,
         message: "Remove action/task denied/unauthorised",
-        value  : JsonNode(nil) )
+        value  : JsonNode(nil) ),
 
-responseMessages["removeError"] = ResponseMessage(
-        code   : "removeError",
-        resCode: Http417,
+        MessageCode.RemoveErrorCode: ResponseMessage(
+        code   : MessageCode.RemoveErrorCode,
+        resCode: StatusCodeExpectationFailed,
         message: "Error removing/deleting information, retry or contact system-admin",
-        value  : JsonNode(nil) )
+        value  : JsonNode(nil) ),
 
-responseMessages["removed"] = ResponseMessage(
-        code   : "removed",
-        resCode: Http200,
+        MessageCode.RemovedCode: ResponseMessage(
+        code   : MessageCode.RemovedCode,
+        resCode: StatusCodeOK,
         message: "Record(s) deleted/removed successfully",
-        value  : JsonNode(nil) )
+        value  : JsonNode(nil) ),
 
-responseMessages["subItems"] = ResponseMessage(
-        code   : "subItems",
-        resCode: Http501,
+        MessageCode.SubItemsCode: ResponseMessage(
+        code   : MessageCode.SubItemsCode,
+        resCode: StatusCodeNotImplemented,
         message: "Record includes sub-items, which must be removed first",
-        value  : JsonNode(nil) )
+        value  : JsonNode(nil) ),
 
-responseMessages["duplicateRec"] = ResponseMessage(
-        code   : "duplicateRec",
-        resCode: Http406,
+        MessageCode.DuplicateRecordCode: ResponseMessage(
+        code   : MessageCode.DuplicateRecordCode,
+        resCode: StatusCodeNotAcceptable,
         message: "Duplicate record exists",
-        value  : JsonNode(nil) )
+        value  : JsonNode(nil) ),
 
-responseMessages["updated"] = ResponseMessage(
-        code   : "updated",
-        resCode: Http200,
+        MessageCode.UpdatedCode: ResponseMessage(
+        code   : MessageCode.UpdatedCode,
+        resCode: StatusCodeOK,
         message: "Information update action completed successfully",
-        value  : JsonNode(nil) )
+        value  : JsonNode(nil) ),
 
-responseMessages["updateError"] = ResponseMessage(
-        code   : "updateError",
-        resCode: Http501,
+        MessageCode.UpdateDeniedCode: ResponseMessage(
+        code   : MessageCode.UpdateDeniedCode,
+        resCode: StatusCodeNotImplemented,
         message: "Error updating information/record(s)",
-        value  : JsonNode(nil) )
+        value  : JsonNode(nil) ),
 
-responseMessages["updateDenied"] = ResponseMessage(
-        code   : "updateDenied",
-        resCode: Http401,
+        MessageCode.UpdateDeniedCode: ResponseMessage(
+        code   : MessageCode.UpdateDeniedCode,
+        resCode: StatusCodeUnAuthorized,
         message: "Update action/task not authorised",
-        value  : JsonNode(nil) )
+        value  : JsonNode(nil) ),
 
-responseMessages["inserted"] = ResponseMessage(
-        code   : "inserted",
-        resCode: Http200,
+        MessageCode.InsertedCode: ResponseMessage(
+        code   : MessageCode.InsertedCode,
+        resCode: StatusCodeOK,
         message: "Information/record(s) inserted/created successfully",
-        value  : JsonNode(nil) )
+        value  : JsonNode(nil) ),
 
-responseMessages["inserError"] = ResponseMessage(
-        code   : "insertError",
-        resCode: Http501,
+        MessageCode.InsertErrorCode: ResponseMessage(
+        code   : MessageCode.InsertErrorCode,
+        resCode: StatusCodeNotImplemented,
         message: "Error inserting/creating new information/record",
-        value  : JsonNode(nil) )
+        value  : JsonNode(nil) ),
 
-responseMessages["recExist"] = ResponseMessage(
-        code   : "recExist",
-        resCode: Http501,
+        MessageCode.RecordExistCode: ResponseMessage(
+        code   : MessageCode.RecordExistCode,
+        resCode: StatusCodeNotImplemented,
         message: "Document/record exists",
+        value  : JsonNode(nil) ),
+
+        MessageCode.TaskErrorCode: ResponseMessage(
+        code   : MessageCode.TaskErrorCode,
+        resCode: StatusCodeOK,
+        message: "Application or Task Specific Error",
+        value  : JsonNode(nil) ),
+
+        MessageCode.UnknownCode: ResponseMessage(
+        code   : MessageCode.UnknownCode,
+        resCode: StatusCodeOK,
+        message: "Unspecified/Unknown response/error message",
         value  : JsonNode(nil) )
 
-responseMessages["unknown"] = ResponseMessage(
-        code   : "unknown",
-        resCode: Http422,
-        message: "Unspecified response/error message",
-        value  : JsonNode(nil) )
+}.toTable