+++ src/mcresponse/response.nim
@@ -14,41 +14,51 @@
 
 import json, httpcore, tables
 import ./messages
+import ./netcode
 
-proc msgFunc*(mcValue: JsonNode; mcCode, mcMsg: string; mcResCode: HttpCode;): ResponseMessage =
+proc msgFunc*(mcCode: MessageCode; mcMsg: string; mcResCode: HttpCode; mcValue: JsonNode;): ResponseMessage =
     result = ResponseMessage(
         code   : mcCode,
         resCode: mcResCode,
         value  : mcValue,
         message: mcMsg)
 
-proc getResMessage*(messageType: string = "unknown";  options: ResponseMessage = ResponseMessage()): ResponseMessage = 
+proc getResMessage*(messageType: MessageCode = MessageCode.UnknownCode;  options: ResponseMessage = ResponseMessage()): ResponseMessage = 
     var
         value: JsonNode
-        code: string
+        code: MessageCode
         resCode: HttpCode
         message: string
 
     # check for the messageType in the standard responseMessages
-    if messageType != "unknown" and responseMessages.hasKey(messageType):
+    if messageType != MessageCode.UnknownCode and responseMessages.hasKey(messageType):
         code = messageType
-    elif options.code != "" and responseMessages.hasKey(options.code):
+    elif responseMessages.hasKey(options.code):
         code = options.code
     else:
         code = messageType
 
     # Obtain the standard response, will always return valid result
     let stdRes =  responseMessages[code]
-
-    message = stdRes.message & " [" & options.message & "]"
-    value = options.value
+    
     resCode = stdRes.resCode
 
-    result = msgFunc(value, code, message, resCode )
+    if options.message != "":
+        message = options.message
+    else:
+        message = stdRes.message
+
+    if options.value != nil:
+        value = options.value
+    else: 
+        value = nil
+    
+
+    result = msgFunc(mcCode = code, mcMsg = message, mcResCode = resCode, mcValue = value )
 
 when isMainModule:
     # standard response message
-    var resMsg = responseMessages["success"]
+    var resMsg = responseMessages[MessageCode.SuccessCode]
     echo "\n"
     echo resMsg
     echo "\n"
@@ -66,4 +76,4 @@ when isMainModule:
 
     var options = ResponseMessage(value: personInfo, message: "personal info")
 
-    echo getResMessage("success", options)
+    echo getResMessage(MessageCode.SuccessCode, options)