+++ tests/testResMessage.nim
@@ -14,10 +14,12 @@ const person2: Person = Person(firstName: "Ola", lastName: "Akindele", age: 200)
 
 let personInfo = %*([person, person2])
 
-var options = ResponseMessage(value: personInfo, message: "personal info")
-var resMsg = getResMessage("success", options)
+var options = ResponseMessage(value: personInfo, message: "Request completed")
+var resMsg = getResMessage(MessageCode.SuccessCode, options)
 echo "\n"
 echo resMsg
+echo "\n"
+echo "msgCode: ", resMsg.code
 echo "\n"
 echo "resCode: ", resMsg.resCode.repr
 echo "\n"
@@ -26,7 +28,7 @@ echo "\n"
 #
 
 test "will response with the correct standard response: ":
-    check resMsg.code == "success"
-    check resMsg.message == "Request completed successfully [personal info]"
+    check resMsg.code.repr == "success"
+    check resMsg.message == "Request completed"
     check resMsg.value == %*([{"firstName":"Abi","lastName":"Akindele","age":100},{"firstName":"Ola","lastName":"Akindele","age":200}])
     check resMsg.resCode.repr == "200"